rename: 사이드바 관련 폴더 이름 변경
