+++ src/components/layout/Sidebar/SidebarNav/index.ts
@@ -1,3 +1,3 @@
 // 1차 사이드바 관련 컴포넌트들을 외부로 내보내기
-export { default as PrimarySidebar } from './PrimarySidebar';
+export { default as PrimarySidebar } from './SidebarNav';
 export { default as SidebarItem } from './SidebarItem';
