+++ src/components/layout/Sidebar/SidebarPanel/SignUpPanel.tsx
