+++ src/components/layout/Sidebar/SidebarPanel/index.ts
@@ -1,3 +1,3 @@
 // 2차 사이드바 관련 컴포넌트들을 외부로 내보내기
-export { default as SecondarySidebar } from './SecondarySidebar';
+export { default as SecondarySidebar } from './SidebarPanel';
 export { default as MenuContent } from './MenuContent';
